network folder created in lib and re arranged some codes
