+++ lib/app/utils/custom_strings.dart
@@ -18,3 +18,13 @@ static const String electronics="electronics";
 static const String mensClothing="men's clothing";
 static const String womensClothing="women's clothing";
 }
+//Api EndPoints
+class MyApiUrl {
+  static String backendUrl = "https://fakestoreapi.com/";
+  static String allProducts = "products";
+  static String allCategories = "products/categories";
+  static String jewelery = "products/category/jewelery";
+  static String electronics = "products/category/electronics";
+  static String mensClothing = "products/category/men's clothing";
+  static String womensClothing = "products/category/women's clothing";
+}
