--- lib/data/repositories/data_repositories.dart
@@ -1,9 +1,0 @@
-class MyApiUrl {
-  static String backendUrl = "https://fakestoreapi.com/";
-  static String allProducts = "products";
-  static String allCategories = "products/categories";
-  static String jewelery = "products/category/jewelery";
-  static String electronics = "products/category/electronics";
-  static String mensClothing = "products/category/men's clothing";
-  static String womensClothing = "products/category/women's clothing";
-}
