+++ lib/presentation/pages/home_screen/controller/home_controller.dart
@@ -2,8 +2,8 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:task/data/data_sources/all_products_request.dart';
 import 'package:task/data/models/onBoard/products/product_model.dart';
+import 'package:task/domain/repositories/home_repo/home_repository.dart';
 
 class HomeController extends GetxController {
   RxList<ProductsModel> productsList = <ProductsModel>[].obs;
@@ -21,7 +21,7 @@ class HomeController extends GetxController {
   void fetchAllitems() async {
     try {
       isLoading.value = true;
-      productsList.value = await NetworkRequester.fetchAllProducts();
+      productsList.value = await HomeRepository.fetchAllProducts();
       productsList.shuffle();
     } finally {
       isLoading.value = false;
